requirements.txt
# Core Framework
fastapi
uvicorn
python-multipart
python-dotenv

# LangChain Ecosystem
langchain
langchain-core
langchain-groq
langchain-community
langchain-text-splitters
langchain-huggingface

# Vector Store & Embeddings
sentence-transformers
faiss-cpu

# Document Processing
pypdf
pdf2image
pytesseract

# Additional Dependencies
pillow
